Change to use Random's default seed of currentTimeMillis()
world.getTime is limited to the ticks in a minecraft day, and thus is very limited for a seed for a random number generator.

An example of an issue this causes is if you set a mine to reset every 20 minutes (one minecraft day) Random will always be seeded with the same value, and thus always produce the same reset mine. (non-fill mode)